Bug#24343330: READ OF OUT-OF-SCOPE (TEMPFILE) IN MYSQL_UPDATE()
The problem was that an IO_CACHE variable used during UPDATE,
was allocated on the stack and later copyied into a heap
allocated variable. However, the pointers inside this struct
was not adjusted, so that they still pointed to the stack.
After the stack frame ended, they became invalid.

This patch fixes the problem by allocating the IO_CACHE on
the heap instead and copying the pointer to this IO_CACHE
struct, rather than copying the contents of the struct.

Issue identified by the new ASAN option
-fsanitize-address-use-after-scope

(cherry picked from commit 4d5ff7bfbf82affbeeedce80a3b9ba92a2302c2c)