Fix bug 82026 / 1596845 (Stack buffer overflow with --ssl-cipher=<more than 4K characters>)
new_VioSSLFd strcats several strings into a 4K-sized buffer. One of
the source strings comes from a --ssl-cipher argument, and has
unbounded length, resulting in buffer overflow if it's too long. Fix
by returning an SSL initialisation error if given ssl-cipher will not
fit into the remaining buffer. Add a debug assert for the case when
ssl-cipher is not given.