Remove O(n) reset_trx_size_updated
Iteration over the whole RW transaction list every time a lock is
released is inefficient. Replace it by a global counter in lock sys,
and individual counters in each transaction. The global counter bump
invalidates all existing transaction dep_size values, and on computing
an updated dep_size value the global counter is copied to the
transaction counter to mark it as valid.

The same need to remove reset_trx_size_updated was independently
identified in MySQL 8.0.3 VATS version, and their implementation was
taken with minor differences.