Update transaction weights after failed CREATE TABLE ... SELECT
An assert fired in lock_remove_all_on_table_for_trx calling
lock_assert_no_rec_locks_in_trx that a transaction still had record
locks after removing some on table to be dropped. This indicates that
the record locks could not be dropped without updating weights and
this is possible by a failing CREATE TABLE ... SELECT statement, which
takes record locks on both the new table records (which are dropped in
the above code path on error) and on the source table records.

Fix by iterating over remaining record locks in the transaction and
decrease weights of their waiters by the number of record locks
removed on the table to be dropped.