Fix negative VATS weights if deadlock checker chooses non-joining trx as victim
Based on a patch by Jiamin Huang: "Fix the problem where negative
weights occur if deadlock happens and a non-joining transacion is
chosen as victim. In that case, a weight update will happen durnig
abort. However, at this point, the weights are not correctly updated
yet, as the weight update for the newly arrived transction won't
happen until after the deadlock check."

The above patch was changed from using an extra field in the trx_t
struct to passing the joiner transaction around in the callstack.
Added a testcase.